microservice/requirements.txt
fastapi[standard]==0.115.6
uvicorn==0.34.0
requests==2.31.0
python-dotenv==1.0.1